fix(adapter): geyser report — liquidType is a byte field, GetValue<int> threw InvalidCastException and killed the GeyserActivated report (the peer never saw the eruption)
The stack was visible in the runtime logs ([Unity:Exception] InvalidCastException
in TrapGeyserPatch.Postfix). Traverse.GetValue<T> requires the exact field
type — byte, not int.

+++ src/CasualtiesUnknownOnline.GameAdapter/Patches/TrapGeyserPatch.cs
@@ -15,5 +15,5 @@ internal static class TrapGeyserPatch
 {
 	private static void Postfix(GeyserScript __instance) =>
 		PatchBridge.Impl?.OnTrapTriggered(EntityEventKind.GeyserActivated, __instance.transform.position,
-			(byte)Traverse.Create(__instance).Field("liquidType").GetValue<int>());
+			Traverse.Create(__instance).Field("liquidType").GetValue<byte>()); // byte, not int — a GetValue<int> cast throws InvalidCastException and kills the report
 }
